# internal/infrastructure/http/apiserver/openapi.yaml
openapi: 3.0.3
info:
  title: Alchemorsel API v3 - Pure Backend
  description: |
    Enterprise-grade recipe management API with AI capabilities - Pure JSON Backend
    
    ## Authentication
    This API uses Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-token>
    ```
    
    ## Rate Limiting
    - Public endpoints: 100 requests per minute
    - Authenticated endpoints: 1000 requests per minute
    
    ## Response Format
    All responses are in JSON format with consistent error structure:
    ```json
    {
      "success": true,
      "data": {...},
      "message": "Success message"
    }
    ```
    
    Error responses:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human readable error message",
        "details": {...}
      }
    }
    ```
  version: 3.0.0
  termsOfService: https://alchemorsel.com/terms
  contact:
    name: API Support
    url: https://alchemorsel.com/support
    email: support@alchemorsel.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.alchemorsel.com/api/v1
    description: Production server
  - url: https://staging-api.alchemorsel.com/api/v1
    description: Staging server
paths:
...
  /ai/generate-recipe:
    post:
      tags:
        - AI Services
      summary: Generate recipe with AI
      description: Use AI to generate a recipe based on ingredients or description
      operationId: generateRecipe
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRecipeRequest'
      responses:
        '200':
          description: Recipe generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedRecipe'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: AI service rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /ai/suggest-ingredients:
    post:
      tags:
        - AI Services
      summary: Suggest ingredients
      description: Get AI suggestions for ingredients based on recipe type or existing ingredients
      operationId: suggestIngredients
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuggestIngredientsRequest'
      responses:
        '200':
          description: Ingredients suggested successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngredientSuggestions'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /ai/analyze-nutrition:
    post:
      tags:
        - AI Services
      summary: Analyze nutrition
      description: Get AI-powered nutritional analysis for ingredients or recipes
      operationId: analyzeNutrition
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NutritionAnalysisRequest'
      responses:
        '200':
          description: Nutrition analysis completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NutritionAnalysis'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter 'Bearer {token}' to authenticate
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Invalid request data"
            details:
              type: object
          required:
            - code
            - message
      required:
        - success
        - error
...
    Ingredient:
      type: object
      properties:
        name:
          type: string
          example: "All-purpose flour"
        amount:
          type: string
          example: "2 1/4"
        unit:
          type: string
          example: "cups"
        notes:
          type: string
          nullable: true
          example: "sifted"
      required:
        - name
        - amount
        - unit
    NutritionInfo:
      type: object
      properties:
        calories:
          type: integer
          example: 150
        protein:
          type: number
          format: float
          example: 2.1
        carbs:
          type: number
          format: float
          example: 21.5
        fat:
          type: number
          format: float
          example: 7.2
        fiber:
          type: number
          format: float
          example: 0.8
        sugar:
          type: number
          format: float
          example: 12.3
        sodium:
          type: number
          format: float
          example: 95.0
...
    GenerateRecipeRequest:
      type: object
      properties:
        prompt:
          type: string
          minLength: 10
          maxLength: 500
          example: "Create a healthy breakfast recipe with oats and berries"
        ingredients:
          type: array
          items:
            type: string
          example: ["oats", "blueberries", "honey", "yogurt"]
        dietary_restrictions:
          type: array
          items:
            type: string
          example: ["vegetarian", "gluten-free"]
        max_time:
          type: integer
          minimum: 5
          maximum: 480
          example: 30
        servings:
          type: integer
          minimum: 1
          maximum: 20
          example: 4
      required:
        - prompt
    GeneratedRecipe:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            title:
              type: string
              example: "Overnight Berry Oat Bowl"
            description:
              type: string
              example: "A nutritious and delicious breakfast bowl"
            ingredients:
              type: array
              items:
                $ref: '#/components/schemas/Ingredient'
            instructions:
              type: array
              items:
                type: string
            prep_time:
              type: integer
              example: 10
            cook_time:
              type: integer
              example: 0
            servings:
              type: integer
              example: 4
            difficulty:
              type: string
              enum: [easy, medium, hard]
              example: "easy"
            category:
              type: string
              example: "Breakfast"
            tags:
              type: array
              items:
                type: string
              example: ["healthy", "breakfast", "oats"]
            nutrition:
              $ref: '#/components/schemas/NutritionInfo'
          required:
            - title
            - description
            - ingredients
            - instructions
            - prep_time
            - cook_time
            - servings
            - difficulty
            - category
        message:
          type: string
          example: "Recipe generated successfully"
      required:
        - success
        - data
        - message
    SuggestIngredientsRequest:
      type: object
      properties:
        existing_ingredients:
          type: array
          items:
            type: string
          example: ["chicken", "onions", "garlic"]
        cuisine:
          type: string
          example: "Italian"
        meal_type:
          type: string
          example: "dinner"
        dietary_restrictions:
          type: array
          items:
            type: string
          example: ["low-carb"]
      required:
        - existing_ingredients
    IngredientSuggestions:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            suggestions:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                    example: "tomatoes"
                  reason:
                    type: string
                    example: "Pairs well with chicken and garlic for Italian cooking"
                  category:
                    type: string
                    example: "vegetables"
                required:
                  - name
                  - reason
                  - category
          required:
            - suggestions
        message:
          type: string
          example: "Ingredients suggested successfully"
      required:
        - success
        - data
        - message
    NutritionAnalysisRequest:
      type: object
      properties:
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/Ingredient'
        servings:
          type: integer
          minimum: 1
          example: 4
      required:
        - ingredients
        - servings
    NutritionAnalysis:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            per_serving:
              $ref: '#/components/schemas/NutritionInfo'
            total:
              $ref: '#/components/schemas/NutritionInfo'
            breakdown:
              type: array
              items:
                type: object
                properties:
                  ingredient:
                    type: string
                    example: "All-purpose flour"
                  nutrition:
                    $ref: '#/components/schemas/NutritionInfo'
                required:
                  - ingredient
                  - nutrition
          required:
            - per_serving
            - total
            - breakdown
        message:
          type: string
          example: "Nutrition analysis completed"
      required:
        - success
        - data
        - message
...